Students can view their grade for the current week
