--- src/vm-angular/src/app/comment/result/result.component.html
@@ -1,18 +1,0 @@
-<div class="container m-3" *ngIf="show_result==true">
-  <!-- when result is clicked we want to see the result for this user for this week -->
-    <div *ngFor="let res of result">
-      {{ res }}
-    </div>
-</div>
-
-  <button class="btn btn-outline-danger btn-block" (click)="toggleResult()">
-    Show Result
-  </button>
-
-<div class="container m-3" *ngIf="show_result==false">
-  <button class="btn btn-outline-danger btn-block" (click)="toggleResult()">
-    Hide Result
-  </button>
-</div>
-
-
