--- src/vm-angular/src/app/mystery/mystery.component.html
@@ -1,3 +1,0 @@
-<p>
-  mystery works!
-</p>
