+++ src/vm-angular/src/app/mystery/mystery.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpService } from '../http.service';
+import {HttpClientModule} from '@angular/common/http';
 
 @Injectable({
   providedIn: 'root'
