+++ src/vm-angular/src/app/mystery/release-view/result-view/result-view-service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
-import { HttpService } from '../../http.service';
+import { HttpService } from '../../../http.service';
+
 
 
 @Injectable({
@@ -7,7 +8,7 @@ import { HttpService } from '../../http.service';
 })
 
 /* Service that allows access to all artifactview component functions dealing with api */
-export class ResultService {
+export class ResultViewService {
 
 
   API_URL = 'http://localhost:8000';
